requirements.txt
Django
requests
pymongo
python-dotenv 